refactor: step1피드백 반영 - PathFinderTest 리팩토링
